Refactor GUI code into its own module and multiple files
